src/sensor-data-forecaster/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
tensorflow==2.17.0
joblib==1.3.2
numpy==1.24.3 